third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-fraction-expected.txt

PASS document.getElementById("testimg").clientWidth==200 is true
PASS successfullyParsed is true

TEST COMPLETE
This test passes if the srcset resource is loaded and displayed as the image
